Fix the "Copy branch name" menu item always being disabled
It seems like the author of the commit 8967faea chose to always disable his
new "Follow this branch" menu item, which he added in that commit. The item
3 is set to "disabled" in both branches of an if statement in the headmenu
proc. Another mistake in the same commit was that he targeted item number 3
of the context menu, which was the "Copy branch name", and not the newly
added "Follow this branch", which had index 4.