Suggest a name for the new branch based on the clipboard contents
When user selects the "Create new branch" menu item, the initial name for
the new branch is taken from the clipboard contents. This allows a user to
move a branch without going through the check out and reset. Instead, the
user is able to "Copy branch name" and then "Create new branch" with the
copied name suggested by default. When confirming the branch creation,
there will be an option to move the existing branch to the new location.

If the clipboard text contains a remote branch name, the
"remotes/remote-name/" prefix is stripped from the suggestion.