Limit the visiblerows output to the fully visible rows
This allows to use this function more meaningfully when determining whether
the currently selected row is visible on screen and, correspondingly, if it
is possible to show the context menu for it. When the selected row is only
partially visible or is just outside the visible region, there is no good
way to position its context menu invoked from the keyboard, and even if
there was, the user would not be able to clearly see what the menu is for.

Remove the weird negative half-row from the setcanvscroll setting of the
ymax for the scroll region size. It's not clear why we want to reduce the
scrollable area so that not all the rows would fit in it.