Fix searching for Windows executables in PATH
Don't append the ".exe" suffix in the _which function in case the command
name already ends with one of the recognized executable file extensions
(the ones listed in the PATHEXT environment variable).

This fixes the problem of the exec function failing to find the executables
in the PATH in case their name already contains the ".exe" (or .bat, .cmd,
etc.) suffix, because _which would effectively search for "name.exe.exe" or
"name.cmd.exe".

The problem was introduced somewhere between Git v2.35.1 and v2.40.